docs: Document usage using autoprogram extension
Integrate the 'sphinxcontrib-autoprogram' extension to automatically
document the CLI.

This module uses 'textwrap.indent', which is only available in Python 3.
While this has been fixed on master [1], it is not part of the most
recent release (0.1.3). While we could blacklist this particular
release, we'd lose code-block formatting of usage samples, which make
this a lot easier to read. As such, it's easier just to stick with
Python 3.

[1] https://bitbucket.org/birkenfeld/sphinx-contrib/commits/97e39e715339c0d6fef22f88a7569a324bc73e66

Signed-off-by: Stephen Finucane <stephen@that.guru>